+++ includes/class/tbs_class.php
@@ -590,7 +590,7 @@ function __construct($Options=null,$VarPrefix='',$FctPrefix='') {
 	} 
 
 	// Set options
-	$this->VarRef =& $GLOBALS;
+	$this->VarRef =$GLOBALS;
 	if (is_array($Options)) $this->SetOption($Options);
 
 	// Links to global variables (cannot be converted to static yet because of compatibility)
@@ -685,7 +685,7 @@ function GetOption($o) {
 
 public function ResetVarRef($ToGlobal) {
 	if ($ToGlobal) {
-		$this->VarRef = &$GLOBALS;
+		$this->VarRef = $GLOBALS;
 	} else {
 		$x = array();
 		$this->VarRef = &$x;